+++ src/ladybug.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <TargetFramework>net5.0</TargetFramework>
 		<PackageId>WelcomeToMonday.Ladybug</PackageId>
-		<Version>2.0.0</Version>
+		<Version>2.1.0</Version>
 		<Authors>M. R. Williams</Authors>
 		<Title>Ladybug Framework</Title>
 		<PackageDescription>A Lightweight Game Framework based on MonoGame</PackageDescription>
